requirements.txt
Flask==2.3.3
Pillow==10.0.0; python_version >= '3.8'
pystray==0.19.4
python-dotenv==1.0.0
